Struct intecture_api::host::remote::Plain

pub struct Plain { /* fields omitted */ }

A Host type that uses an unencrypted socket.

Warning! An unencrypted host is susceptible to eavesdropping and MITM attacks, and should only be used on secure private networks.

Methods

impl Plain

pub fn connect(
addr: &str,
handle: &Handle
) -> Box<Future<Item = Plain, Error = Error>>

Create a new Host connected to the given address.