Struct intecture_api::package::Package

pub struct Package<H: Host> { /* fields omitted */ }

Represents a system package to be managed for a host.

Example

Install a package and print the result.

extern crate futures;
extern crate intecture_api;
extern crate tokio_core;

use futures::{future, Future};
use intecture_api::errors::*;
use intecture_api::prelude::*;
use tokio_core::reactor::Core;

let mut core = Core::new().unwrap();
let handle = core.handle();

let host = Local::new(&handle).wait().unwrap();

let nginx = Package::new(&host, "nginx");
let result = nginx.install().and_then(|status| {
    match status {
        // We're performing the install
        Some(status) => Box::new(status.result().unwrap()
            .map(|_| println!("Installed"))
            .map_err(|e| {
                match *e.kind() {
                    ErrorKind::Command(ref output) => println!("Failed with output: {}", output),
                    _ => unreachable!(),
                }
                e
            })),

        // This package is already installed
        None => {
            println!("Already installed");
            Box::new(future::ok(())) as Box<Future<Item = _, Error = Error>>
        },
    }
});

core.run(result).unwrap();

Methods

impl<H: Host + 'static> Package<H>

pub fn new(host: &H, name: &str) -> Package<H>

Create a new Package with the default Provider.

pub fn with_provider(host: &H, provider: Provider, name: &str) -> Package<H>

Create a new Package with the specified Provider.

Example

extern crate futures;
extern crate intecture_api;
extern crate tokio_core;

use futures::Future;
use intecture_api::package::Provider;
use intecture_api::prelude::*;
use tokio_core::reactor::Core;

let mut core = Core::new().unwrap();
let handle = core.handle();

let host = Local::new(&handle).wait().unwrap();

Package::with_provider(&host, Provider::Yum, "nginx");

pub fn installed(&self) -> Box<Future<Item = bool, Error = Error>>

Check if the package is installed.

pub fn install(
&self
) -> Box<Future<Item = Option<CommandStatus>, Error = Error>>

Install the package.

Idempotence

This function is idempotent, which is represented by the type Future<Item = Option<..>, ...>. Thus if it returns Option::None then the package is already installed, and if it returns Option::Some then Intecture is attempting to install the package.

If this fn returns Option::Some<..>, the nested tuple will hold handles to the live output and the result of the installation. Under the hood this reuses the Command endpoint, so see Command docs for detailed usage.

pub fn uninstall(
&self
) -> Box<Future<Item = Option<CommandStatus>, Error = Error>>

Uninstall the package.

Idempotence

This function is idempotent, which is represented by the type Future<Item = Option<..>, ...>. Thus if it returns Option::None then the package is already uninstalled, and if it returns Option::Some then Intecture is attempting to uninstall the package.

If this fn returns Option::Some<..>, the nested tuple will hold handles to the live output and the result of the deinstallation. Under the hood this reuses the Command endpoint, so see Command docs for detailed usage.