Struct intecture_api::service::Service

pub struct Service<H: Host> { /* fields omitted */ }

Represents a system service to be managed for a host.

Example

Enable and start a service.

extern crate futures;
extern crate intecture_api;
extern crate tokio_core;

use futures::{future, Future};
use intecture_api::errors::Error;
use intecture_api::prelude::*;
use tokio_core::reactor::Core;

let mut core = Core::new().unwrap();
let handle = core.handle();

let host = Local::new(&handle).wait().unwrap();

let nginx = Service::new(&host, "nginx");
let result = nginx.enable()
    .and_then(|_| {
        nginx.action("start")
            .and_then(|maybe_status| {
                match maybe_status {
                    Some(status) => Box::new(status.result().unwrap().map(|_| ())) as Box<Future<Item = (), Error = Error>>,
                    None => Box::new(future::ok(())),
                }
            })
    });

core.run(result).unwrap();

Methods

impl<H: Host + 'static> Service<H>

pub fn new(host: &H, name: &str) -> Service<H>

Create a new Service with the default Provider.

pub fn with_provider(host: &H, provider: Provider, name: &str) -> Service<H>

Create a new Service with the specified Provider.

Example

extern crate futures;
extern crate intecture_api;
extern crate tokio_core;

use futures::Future;
use intecture_api::service::Provider;
use intecture_api::prelude::*;
use tokio_core::reactor::Core;

let mut core = Core::new().unwrap();
let handle = core.handle();

let host = Local::new(&handle).wait().unwrap();

Service::with_provider(&host, Provider::Systemd, "nginx");

pub fn running(&self) -> Box<Future<Item = bool, Error = Error>>

Check if the service is currently running.

pub fn action(
&self,
action: &str
) -> Box<Future<Item = Option<CommandStatus>, Error = Error>>

Perform an action for the service, e.g. "start".

Cross-platform services

By design, actions are specific to a particular service and are not cross-platform. Actions are defined by the package maintainer that wrote the service configuration, thus users should take care that they adhere to the configuration for each platform they target.

Idempotence

This function is idempotent when running either the "start" or "stop" actions, as it will check first whether the service is already running. Idempotence is represented by the type Future<Item = Option<..>, ...>. Thus if it returns Option::None then the service is already in the required state, and if it returns Option::Some then Intecture is attempting to transition the service to the required state.

If this fn returns Option::Some<..>, the nested tuple will hold handles to the live output and the result of the action. Under the hood this reuses the Command endpoint, so see Command docs for detailed usage.

pub fn enabled(&self) -> Box<Future<Item = bool, Error = Error>>

Check if the service will start at boot.

pub fn enable(&self) -> Box<Future<Item = Option<()>, Error = Error>>

Instruct the service to start at boot.

Idempotence

This function is idempotent, which is represented by the type Future<Item = Option<..>, ...>. Thus if it returns Option::None then the service is already enabled, and if it returns Option::Some then Intecture is attempting to enable the service.

If this fn returns Option::Some<..>, the nested tuple will hold handles to the live output and the 'enable' command result. Under the hood this reuses the Command endpoint, so see Command docs for detailed usage.

pub fn disable(&self) -> Box<Future<Item = Option<()>, Error = Error>>

Prevent the service from starting at boot.

Idempotence

This function is idempotent, which is represented by the type Future<Item = Option<..>, ...>. Thus if it returns Option::None then the service is already disabled, and if it returns Option::Some then Intecture is attempting to disable the service.

If this fn returns Option::Some<..>, the nested tuple will hold handles to the live output and the 'disable' command result. Under the hood this reuses the Command endpoint, so see Command docs for detailed usage.