Module intecture_api::telemetry

System generated data about your host.

Telemetry is retrieved automatically when you create a new Host, which is nice of it. Call Host.telemetry() to access it.

Structs

Cpu
Information about the Hosts CPU.

FsMount
Information about a specific filesystem mount.

Os
Information about the Hosts OS.

Telemetry
Top level structure that contains static information about a Host.

User
Information on the current user

Enums

LinuxDistro
Linux distribution name

OsFamily
Operating system family

OsPlatform
Operating system name